Ajout restrictions Admin

+++ app/Http/Controllers/ProfileController.php
@@ -97,7 +97,7 @@ class ProfileController extends Controller
      */
     public function destroy(User $user)
     {
-        /* $this->authorize ('manage', $user); */
+        $this->authorize('manage', $user);
         $user->delete();
 
         return response()->json($user);
